pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-search-mcp"
version = "0.1.0"
description = "MCP server for searching git commit diffs with regex"
requires-python = ">=3.13"
dependencies = [
    "mcp>=0.9.0",
    "GitPython>=3.1.0",
    "pydantic>=2.0.0"
]

[project.scripts]
git-search-mcp = "git_search_mcp.server:main"

[tool.uv]
dev-dependencies = [
    "pre-commit>=3.0.0",
    "ruff>=0.1.0"
]